Rename temporary_worktree
